Refactor: file and imports (tests)
