+++ test/filter_test.go
@@ -3,7 +3,7 @@ package rimindertest
 import (
 	"testing"
 
-	"github.com/Xalrandion/go-riminder-api/riminder"
+	"github.com/Xalrandion/go-riminder-api"
 )
 
 func TestFilterListNoError(t *testing.T) {
